backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.49
authlib==1.3.0
httpx==0.27.0
python-dotenv==1.0.1
orgparse==0.4.20231004
gitpython==3.1.43
pytest==8.2.0
itsdangerous==2.2.0
starlette==0.37.2
pytest-asyncio==0.23.6
